Refactored to use headers + added cmake/make building (with helper bash script)
