+++ .clangd
@@ -1,0 +1,2 @@
+CompileFlags:
+    Add: [-std=c++20]
