+++ src/argparse.cpp
@@ -1,17 +1,8 @@
-#include <map>
-#include <string>
-#include <vector>
+#include "include/argparse.hpp"
+
 
 using namespace std;
 
-
-/**
-* @brief Parse arguments into a map of strings. Based on assumption of 1 arg per flag
-*
-* @param nargs Number of arguments
-* @param args Arguments
-* @returns map of arguments by flags. Both flag and arg as strings
-*/
 map<string, string> parse_args(int nargs, const char* args[]){
     map<string, string> arguments;
     string current_flag;
@@ -30,13 +21,6 @@ map<string, string> parse_args(int nargs, const char* args[]){
     return arguments;
 }
 
-/**
-* @brief Check if parsed arguments contain a given flag
-*
-* @param args Arguments
-* @param flag Flag to check for
-* @returns bool True if the flag exists
-*/
 bool check_flag(map<string, string> args, string flag){
     for (auto const& [key, val] : args){
         if(key == flag){
@@ -46,13 +30,6 @@ bool check_flag(map<string, string> args, string flag){
     return false;
 }
 
-/**
-* @brief Check if parsed arguments contain all of the required flags
-*
-* @param args Arguments
-* @param required Vector of flags to check for
-* @returns bool True if all flags exist
-*/
 bool check_required(map<string, string> args, vector<string> required){
     for(int i=0;i<required.size();i++){
         if(!check_flag(args, required.at(i))){
